# contracts/openapi/manoksha-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Manoksha Collections API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "RefreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reseller/wallet": {
      "get": {
        "tags": [
          "Reseller portal"
        ],
        "operationId": "ResellerWallet",
        "parameters": [
          {
            "name": "beforeSeq",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reseller/wallet/deposits": {
      "get": {
        "tags": [
          "Reseller portal"
        ],
        "operationId": "ResellerDeposits",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepositDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reseller portal"
        ],
        "operationId": "SubmitDeposit",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "amount",
                  "method",
                  "reference"
                ],
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "format": "double"
                  },
                  "method": {
                    "type": "string"
                  },
                  "reference": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string"
                  },
                  "proof": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "amount": {
                  "style": "form"
                },
                "method": {
                  "style": "form"
                },
                "reference": {
                  "style": "form"
                },
                "note": {
                  "style": "form"
                },
                "proof": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reseller/wallet/deposits/online": {
      "post": {
        "tags": [
          "Reseller portal"
        ],
        "operationId": "StartOnlineDeposit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartOnlineDepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlineDepositDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Reseller portal"
        ],
        "operationId": "ResellerOnlineDeposits",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OnlineDepositDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reseller/wallet/deposits/online/{id}": {
      "get": {
        "tags": [
          "Reseller portal"
        ],
        "operationId": "ResellerOnlineDeposit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlineDepositDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reseller/wallet/deposits/{id}/proof": {
      "get": {
        "tags": [
          "Reseller portal"
        ],
        "operationId": "ResellerDepositProof",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/reseller/checkout": {
      "post": {
        "tags": [
          "Reseller portal"
        ],
        "operationId": "ResellerCheckout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResellerCheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/wallet/resellers/{resellerId}/ledger": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "operationId": "GetResellerLedger",
        "parameters": [
          {
            "name": "resellerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "beforeSeq",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/wallet/resellers/{resellerId}/adjustments": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "operationId": "AdjustResellerWallet",
        "parameters": [
          {
            "name": "resellerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualAdjustmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerEntryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/wallet/deposits": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "operationId": "ListDeposits",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resellerId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepositDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/wallet/deposits/{id}/approve": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "operationId": "ApproveDeposit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/wallet/deposits/{id}/reject": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "operationId": "RejectDeposit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectDepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/wallet/deposits/{id}/proof": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "operationId": "GetDepositProof",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CheckoutLineRequest": {
        "type": "object",
        "properties": {
          "skuId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CheckoutResult": {
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "$ref": "#/components/schemas/OrderDto"
          },
          "walletBalance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "inquiry": {
            "$ref": "#/components/schemas/InquiryDto"
          }
        },
        "additionalProperties": false
      },
...
      "DeliveryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "mobile": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "addressLine": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "pin": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DepositDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "resellerId": {
            "type": "string",
            "format": "uuid"
          },
          "resellerNumber": {
            "type": "string",
            "nullable": true
          },
          "resellerName": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "method": {
            "type": "string",
            "nullable": true
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "resellerNote": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reviewNote": {
            "type": "string",
            "nullable": true
          },
          "ledgerEntryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "InquiryDto": {
        "type": "object",
        "properties": {
          "reference": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "whatsAppUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LedgerEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "seq": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "direction": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "balanceBefore": {
            "type": "number",
            "format": "double"
          },
          "balanceAfter": {
            "type": "number",
            "format": "double"
          },
          "orderId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "orderNumber": {
            "type": "string",
            "nullable": true
          },
          "depositRequestId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "reversesEntryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LedgerPage": {
        "type": "object",
        "properties": {
          "balance": {
            "type": "number",
            "format": "double"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerEntryDto"
            },
            "nullable": true
          },
          "nextBeforeSeq": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ManualAdjustmentRequest": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OnlineDepositDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "providerPaymentRef": {
            "type": "string",
            "nullable": true
          },
          "payment": {
            "$ref": "#/components/schemas/OnlineDepositPaymentDto"
          }
        },
        "additionalProperties": false
      },
      "OnlineDepositPaymentDto": {
        "type": "object",
        "properties": {
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "redirectUrl": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "OrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "channel": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "resellerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "fulfillmentBranchId": {
            "type": "string",
            "format": "uuid"
          },
          "fulfillmentBranchName": {
            "type": "string",
            "nullable": true
          },
          "delivery": {
            "$ref": "#/components/schemas/DeliveryDto"
          },
          "merchandiseTotal": {
            "type": "number",
            "format": "double"
          },
          "shippingFee": {
            "type": "number",
            "format": "double"
          },
          "grandTotal": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderLineDto"
            },
            "nullable": true
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderStatusChangeDto"
            },
            "nullable": true
          },
          "helpWhatsAppUrl": {
            "type": "string",
            "nullable": true
          },
          "costOfGoods": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderLineDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "skuId": {
            "type": "string",
            "format": "uuid"
          },
          "skuCode": {
            "type": "string",
            "nullable": true
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "variantName": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "retailUnitPrice": {
            "type": "number",
            "format": "double"
          },
          "discountSource": {
            "type": "string",
            "nullable": true
          },
          "discountPct": {
            "type": "number",
            "format": "double"
          },
          "discountAmountPerUnit": {
            "type": "number",
            "format": "double"
          },
          "finalUnitPrice": {
            "type": "number",
            "format": "double"
          },
          "lineTotal": {
            "type": "number",
            "format": "double"
          },
          "commercialTermVersion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OrderStatusChangeDto": {
        "type": "object",
        "properties": {
          "fromStatus": {
            "type": "string",
            "nullable": true
          },
          "toStatus": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RejectDepositRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ResellerCheckoutRequest": {
        "type": "object",
        "properties": {
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckoutLineRequest"
            },
            "nullable": true
          },
          "resellerCustomerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "delivery": {
            "$ref": "#/components/schemas/DeliveryDto"
          },
          "saveCustomer": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ReviewDepositRequest": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StartOnlineDepositRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "TokenPair": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          },
          "refreshTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}